+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Remove empty lines in drop-down menus in Settings > Orders > Number generator options. Refs UIU-3424.
 * AccountDetailsContainer - Show loading while data is loading to prevent page crash when visiting the Fee/fines details page via URL. Fixes UIU-3428.
 * PermissionsModal to search in permissionName. Refs UIU-3436.
+* Display custom fields in different accordions depending on the selected section. Refs UIU-3131.
 
 ## [12.1.7] (https://github.com/folio-org/ui-users/tree/v12.1.7) (2025-06-03)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v12.1.6...v12.1.7)